Move muxConnectManager to its own file
